backend/pyproject.toml
[project]
name = "margin-collateral-agent"
version = "0.2.0"
description = "AI-powered agent for OTC derivatives collateral management under ISDA/CSA agreements"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.1",
    "python-dateutil>=2.8.0",
    "landingai>=0.3.0",
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",           # Testing framework
    "pytest-cov>=4.0.0",       # Test coverage
    "pytest-asyncio>=0.21.0",  # Async test support
    "black>=24.0.0",           # Code formatting
    "ruff>=0.1.0",             # Fast linter
    "mypy>=1.0.0",             # Type checking
]
